+++ src/jsMain/kotlin/at/posselt/pfrpg2e/camping/CampingUpdateBuilder.kt
@@ -10,12 +10,32 @@ import js.objects.unsafeJso
 
 @DocumentUpdateDsl
 @Suppress("unused")
+class CookingResultPropertyBuilder(basePath: String, updates: Record<String, Any?>, propertyName: String) :
+    RecordPropertyUpdateBuilder<Record<String, CookingResult>>(basePath, updates, propertyName) {
+    val result = PropertyUpdateBuilder<String?>(propertyName, updates, "result")
+    val skill = PropertyUpdateBuilder<String>(propertyName, updates, "skill")
+
+    operator fun invoke(action: CookingResultPropertyBuilder.() -> Unit) = action()
+}
+
+@Suppress("unused")
+class ActorMealPropertyBuilder(basePath: String, updates: Record<String, Any?>, propertyName: String) :
+    RecordPropertyUpdateBuilder<Record<String, ActorMeal>>(basePath, updates, propertyName) {
+    val actorUuid = PropertyUpdateBuilder<String>(propertyName, updates, "actorUuid")
+    val favoriteMeal = PropertyUpdateBuilder<String?>(propertyName, updates, "favoriteMeal")
+    val chosenMeal = PropertyUpdateBuilder<String>(propertyName, updates, "chosenMeal")
+
+    operator fun invoke(action: ActorMealPropertyBuilder.() -> Unit) = action()
+}
+
+@DocumentUpdateDsl
+@Suppress("unused")
 class CookingPropertyBuilder(basePath: String, updates: Record<String, Any?>, propertyName: String) :
     RecordPropertyUpdateBuilder<Cooking>(basePath, updates, propertyName) {
-    val actorMeals = PropertyUpdateBuilder<Array<ActorMeal>>(propertyName, updates, "actorMeals")
+    val actorMeals = ActorMealPropertyBuilder(propertyName, updates, "actorMeals")
     val knownRecipes = PropertyUpdateBuilder<Array<String>>(propertyName, updates, "knownRecipes")
     val homebrewMeals = PropertyUpdateBuilder<Array<RecipeData>>(propertyName, updates, "homebrewMeals")
-    val results = PropertyUpdateBuilder<Array<CookingResult>>(propertyName, updates, "results")
+    val results = CookingResultPropertyBuilder(propertyName, updates, "results")
     val minimumSubsistence = PropertyUpdateBuilder<Int>(propertyName, updates, "minimumSubsistence")
 
     operator fun invoke(action: CookingPropertyBuilder.() -> Unit) = action()
@@ -59,7 +79,7 @@ class RestSettingsPropertyBuilder(basePath: String, updates: Record<String, Any?
 class CampingUpdateBuilder(val updates: Record<String, Any?>, basePath: String = "") {
     val actorUuids = PropertyUpdateBuilder<Array<String>>(basePath, updates, "actorUuids")
     val campingActivities =
-        RecordPropertyUpdateBuilder<Record<String, CampingActivityWithId>>(basePath, updates, "campingActivities")
+        RecordPropertyUpdateBuilder<Record<String, CampingActivity>>(basePath, updates, "campingActivities")
     val cooking = CookingPropertyBuilder(basePath, updates, "cooking")
     val currentRegion = PropertyUpdateBuilder<String>(basePath, updates, "currentRegion")
     val homebrewCampingActivities =
@@ -86,6 +106,8 @@ class CampingUpdateBuilder(val updates: Record<String, Any?>, basePath: String =
     val secondsSpentHexploring = PropertyUpdateBuilder<Int>(basePath, updates, "secondsSpentHexploring")
     val resetTimeTrackingAfterOneDay = PropertyUpdateBuilder<Boolean>(basePath, updates, "resetTimeTrackingAfterOneDay")
     val travelModeActive = PropertyUpdateBuilder<Boolean>(basePath, updates, "travelModeActive")
+    val secondsSpentForcedMarching = PropertyUpdateBuilder<Int>(basePath, updates, "secondsSpentForcedMarching")
+    val forcedMarchActive = PropertyUpdateBuilder<Boolean>(basePath, updates, "forcedMarchActive")
     val restingTrack = RestingTrackPropertyBuilder(basePath, updates, "restingTrack")
     val regionSettings = RegionSettingsPropertyBuilder(basePath, updates, "regionSettings")
     val restSettings = RestSettingsPropertyBuilder(basePath, updates, "restSettings")