+++ src/jsMain/kotlin/at/posselt/pfrpg2e/camping/MealDiffing.kt
@@ -1,70 +1,50 @@
 package at.posselt.pfrpg2e.camping
 
-import at.posselt.pfrpg2e.Config
 import at.posselt.pfrpg2e.data.checks.DegreeOfSuccess
 import at.posselt.pfrpg2e.fromCamelCase
 import at.posselt.pfrpg2e.takeIfInstance
-import at.posselt.pfrpg2e.utils.asAnyObjectList
+import at.posselt.pfrpg2e.utils.asAnyObject
+import at.posselt.pfrpg2e.utils.asSequence
 import at.posselt.pfrpg2e.utils.buildPromise
+import at.posselt.pfrpg2e.utils.parseChanges
 import at.posselt.pfrpg2e.utils.postChatTemplate
 import at.posselt.pfrpg2e.utils.t
+import at.posselt.pfrpg2e.utils.toMap
 import com.foundryvtt.core.AnyObject
 import com.foundryvtt.core.documents.Actor
 import com.foundryvtt.core.documents.onPreUpdateActor
 import com.foundryvtt.core.helpers.TypedHooks
-import com.foundryvtt.core.utils.diffObject
 import com.foundryvtt.core.utils.getProperty
-import js.objects.Object
+import js.array.component1
+import js.array.component2
+import js.objects.Record
 import js.objects.recordOf
-
-private const val mealResultPath = "flags.${Config.moduleId}.camping-sheet.cooking.results"
-
-private fun doObjectArraysDiffer(source: List<AnyObject>, target: List<AnyObject>): Boolean {
-    return source.size != target.size ||
-            (source.asSequence() zip target.asSequence())
-                .any { (first, second) ->
-                    Object.keys(diffObject(first, second)).isNotEmpty()
-                }
-}
-
-private fun relevantUpdate(camping: CampingData, update: AnyObject): Boolean {
-    val current = camping.cooking.results
-        .sortedBy { it.recipeId }
-    val updated = getProperty(update, mealResultPath)
-        ?.unsafeCast<Array<CookingResult>>()
-        ?.sortedBy { it.recipeId }
-        ?: emptyList()
-    return doObjectArraysDiffer(current.asAnyObjectList(), updated.asAnyObjectList())
-}
 
 private suspend fun checkPreActorMealUpdate(actor: Actor, update: AnyObject) {
     val camping = actor.takeIfInstance<CampingActor>()?.getCamping() ?: return
-    if (!relevantUpdate(camping, update)) return
-    console.log("Received camping update", update)
+    val campingUpdate = getProperty(update, campingPath).unsafeCast<AnyObject?>() ?: return
+    val changes = parseChanges(camping.asAnyObject(), campingUpdate, setOf("results"), "cooking") ?: return
+    val applied = getProperty(changes.applied, "cooking.results").unsafeCast<Record<String, CookingResult>?>() ?: return
+    console.log("Received camping update", update, applied)
     val recipesById = camping.getAllRecipes().associateBy { it.id }
-    val current = camping.cooking.results.associateBy { it.recipeId }
-    val updated = getProperty(update, mealResultPath)
-        ?.unsafeCast<Array<CookingResult>>()
-        ?: emptyArray()
-    updated
-        .filter {
-            it.result != null && it.result != current[it.recipeId]?.result
-        }
-        .forEach { result ->
-            val degree = result.result?.let { fromCamelCase<DegreeOfSuccess>(it) }
-            val data = recipesById[result.recipeId]
+    val current = camping.cooking.results.toMap()
+    applied.asSequence()
+        .filter { (recipeId, data) -> data.result != null && data.result != current[recipeId]?.result }
+        .forEach { (recipeId, data) ->
+            val degree = data.result?.let { fromCamelCase<DegreeOfSuccess>(it) }
+            val recipeData = recipesById[recipeId]
             val message = when (degree) {
-                DegreeOfSuccess.CRITICAL_FAILURE -> data?.criticalFailure?.message
-                DegreeOfSuccess.SUCCESS -> data?.success?.message
-                DegreeOfSuccess.CRITICAL_SUCCESS -> data?.criticalSuccess?.message
+                DegreeOfSuccess.CRITICAL_FAILURE -> recipeData?.criticalFailure?.message
+                DegreeOfSuccess.SUCCESS -> recipeData?.success?.message
+                DegreeOfSuccess.CRITICAL_SUCCESS -> recipeData?.criticalSuccess?.message
                 else -> null
             }
             postChatTemplate(
                 templatePath = "chatmessages/apply-meal-result.hbs",
                 templateContext = recordOf(
-                    "id" to result.recipeId,
-                    "name" to data?.name,
-                    "degree" to result.result,
+                    "id" to recipeId,
+                    "name" to recipeData?.name,
+                    "degree" to data.result,
                     "label" to degree?.let { t(it) },
                     "message" to message,
                     "failure" to (degree == DegreeOfSuccess.FAILURE),