+++ src/jsMain/kotlin/at/posselt/pfrpg2e/migrations/migrations/Migration20.kt
@@ -1,9 +1,13 @@
 package at.posselt.pfrpg2e.migrations.migrations
 
+import at.posselt.pfrpg2e.camping.ActorMeal
 import at.posselt.pfrpg2e.camping.CampingActivity
 import at.posselt.pfrpg2e.camping.CampingData
+import at.posselt.pfrpg2e.camping.CookingResult
 import at.posselt.pfrpg2e.utils.toMutableRecord
 import com.foundryvtt.core.Game
+import com.foundryvtt.core.utils.fromUuid
+import kotlinx.coroutines.await
 import kotlinx.js.JsPlainObject
 
 @JsPlainObject
@@ -14,6 +18,19 @@ private external interface OldCampingActivity {
     var selectedSkill: String?
 }
 
+@JsPlainObject
+private external interface OldCookingResult {
+    val recipeId: String
+    var result: String?
+    val skill: String
+}
+
+@JsPlainObject
+external interface OldActorMeal {
+    var actorUuid: String
+    var favoriteMeal: String?
+    var chosenMeal: String
+}
 
 class Migration20 : Migration(20) {
     override suspend fun migrateCamping(game: Game, camping: CampingData) {
@@ -26,5 +43,25 @@ class Migration20 : Migration(20) {
                 )
             }
             .toMutableRecord()
+        camping.cooking.results = camping.cooking.results.unsafeCast<Array<OldCookingResult>>()
+            .map {
+                it.recipeId to CookingResult(
+                    result = it.result,
+                    skill = it.skill,
+                )
+            }
+            .toMutableRecord()
+        camping.cooking.actorMeals = camping.cooking.actorMeals.unsafeCast<Array<OldActorMeal>>()
+            .mapNotNull {
+                val id = fromUuid(it.actorUuid).await()?.id ?: return@mapNotNull null
+                id to ActorMeal(
+                    actorUuid = it.actorUuid,
+                    favoriteMeal = it.favoriteMeal,
+                    chosenMeal = it.chosenMeal,
+                )
+            }
+            .toMutableRecord()
+        camping.forcedMarchActive = false
+        camping.secondsSpentForcedMarching = 0
     }
 }