+++ src/jsTest/kotlin/at/posselt/pfrpg2e/camping/CampingUpdateBuilderTest.kt
@@ -33,18 +33,18 @@ class CampingUpdateBuilderTest {
     @Test
     fun testActorMeals() {
         val expected = recordOf(
-            "cooking.actorMeals" to arrayOf(
-                ActorMeal(
-                    actorUuid = "actor",
+            "cooking.actorMeals" to recordOf(
+                "actor" to ActorMeal(
+                    actorUuid = "actoruuid",
                     chosenMeal = "meal",
                 )
             )
         )
         val result = buildCampingUpdate {
             cooking.actorMeals.set(
-                arrayOf(
-                    ActorMeal(
-                        actorUuid = "actor",
+                recordOf(
+                    "actor" to ActorMeal(
+                        actorUuid = "actoruuid",
                         chosenMeal = "meal",
                     )
                 )
@@ -56,9 +56,9 @@ class CampingUpdateBuilderTest {
     @Test
     fun testNestedCooking() {
         val expected = recordOf(
-            "cooking.actorMeals" to arrayOf(
-                ActorMeal(
-                    actorUuid = "actor",
+            "cooking.actorMeals" to recordOf(
+                "actor" to ActorMeal(
+                    actorUuid = "actoruuid",
                     chosenMeal = "meal",
                 )
             )
@@ -66,9 +66,9 @@ class CampingUpdateBuilderTest {
         val result = buildCampingUpdate {
             cooking {
                 actorMeals.set(
-                    arrayOf(
-                        ActorMeal(
-                            actorUuid = "actor",
+                    recordOf(
+                        "actor" to ActorMeal(
+                            actorUuid = "actoruuid",
                             chosenMeal = "meal",
                         )
                     )
@@ -86,9 +86,9 @@ class CampingUpdateBuilderTest {
         val result = buildCampingUpdate {
             cooking {
                 actorMeals.set(
-                    arrayOf(
-                        ActorMeal(
-                            actorUuid = "actor",
+                    recordOf(
+                        "actor" to ActorMeal(
+                            actorUuid = "actoruuid",
                             chosenMeal = "meal",
                         )
                     )